+++ src/main/kotlin/io/telegramkt/model/contact/Contact.kt
@@ -8,6 +8,6 @@ data class Contact(
     @SerialName("phone_number") val phoneNumber: String,
     @SerialName("first_name") val firstName: String,
     @SerialName("last_name") val lastName: String? = null,
-    @SerialName("user_id") val userId: Int? = null,
+    @SerialName("user_id") val userId: Long? = null,
     @SerialName("vcard") val vCard: String? = null,
 )
